Working with hfst-preprocessing, testing first for sje.
Reintroduced abbreviations in sje for testing (it may be removed again ad lib, 
since abbrs are not needed in sje, for now).

I did not use the flag diacritic option (as this one was geared towards
distinguishing the dot from the nodot, whereas this was not the
decisive factor, rather the decisive factor was TRAB vs. ITRAB. Now,
all periods are separate readings, with two possible analyses:
+CLB and +SENT+CLB. The former is for non-boundary-periods, and
the latter is for periods also being bondaries. The abbreviations are
then tabbed for type, and a simple cg rule removes the SENT boundary
when TRAB to the left. Subsequent CG rules will then constitute
the search domain according to whether SENT is present or not.

Next step is sme, where we can test on a larger scale.


git-svn-id: https://gtsvn.uit.no/langtech/trunk/langs/sje@166155 c7155fb1-f0a7-4240-a2fc-2600b6f42f90